+++ R/app_server.R
@@ -31,12 +31,16 @@ app_server = function(input, output, session) {
       "border-radius:8px;",
       "font-family:\"Courier New\",monospace;"
     )
+    w = shinybrowser::get_width() / 72
+    h = (1800 - 40) / 72
     suppressWarnings(ggiraph::girafe(ggobj = g,
+                                     width_svg = w,
+                                     height_svg = h,
                     options = list(
                       ggiraph::opts_selection(
                         type = "single"),
-                      ggiraph::opts_sizing(
-                        width = 0.8),
+                      ggiraph::opts_sizing(rescale = FALSE),
+                      ggiraph::opts_zoom(max = 5),
                       ggiraph::opts_tooltip(
                         css = tooltip_css,
                         use_fill = FALSE)