+++ R/utils.R
@@ -1,13 +1,19 @@
+#' function to return treeview options
+available_treeview = function() {
+  all_trees = list.files(system.file("app", "www", "data", "treeview",
+                                     package = "tfpbrowser"), pattern = "\\.rds$")
+  names(all_trees) = all_trees %>%
+    stringr::str_replace_all("_|-|\\.rds", " ") %>%
+    stringr::str_trim() %>%
+    stringr::str_to_title()
+  return(all_trees)
+  }
+
 #' function to return folder name
-#' @param type Character string detailing type of widget to show
+#' @param type Choice of treeview widget to show
 get_filename = function(type) {
-  filename = switch(type,
-                    "Logistic growth rate" = "tree-logistic_growth_rate.rds", # nolint
-                    "Simple logistic growth rate" = "tree-simple_logistic_growth_rate.rds", # nolint
-                    "Simple trait log odds" = "tree-sim_trait_logodds.rds" # nolint
-  )
   filename = system.file("app", "www", "data", "treeview",
-                         filename,
+                         type,
                          package = "tfpbrowser",
                          mustWork = TRUE)
   return(filename)
@@ -54,29 +60,10 @@ get_all_files = function(cluster_choice) {
 #' function to tidy up table output
 #' @param table_to_display Data frame or tibble containing messy outputs
 reformat_table = function(table_to_display) {
-  if (nrow(table_to_display) == 1) {
-    output = table_to_display[, -1]$x
-    if (!is.na(output)) {
-      output = output %>%
-        stringr::str_split(pattern = "\n") %>%
-        unlist() %>%
-        stringr::str_trim() %>%
-        tibble::as_tibble() %>%
-        tidyr::separate(.data$value,
-                        into = c("x", "y"),
-                        sep = "  ",
-                        extra = "merge") %>%
-        dplyr::mutate(x = stringr::str_trim(.data$x),
-                      y = stringr::str_trim(.data$y)) %>%
-        `colnames<-`(.[1, ]) %>% # nolint
-        dplyr::slice(-1)
-    } else {
-      output = tibble::tibble(x = "Nothing to display")
-    }
-  } else {
-    output = janitor::clean_names(table_to_display,
-                                  case = "title")
+  if (all(table_to_display[[1]] == seq_len(nrow(table_to_display)))) {
+    table_to_display = table_to_display[, -1]
   }
+  output = janitor::clean_names(table_to_display, case = "title")
   return(output)
 }